backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic[email]==2.10.3
supabase==2.10.0
python-multipart==0.0.20
